app/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiogram-forms==1.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beanie==1.25.0
certifi==2024.2.2
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
lazy-model==0.2.0
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
motor==3.4.0
multidict==6.0.5
orjson==3.9.15
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.14.6
Pygments==2.17.2
pymongo==4.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
toml==0.10.2
traitlets==5.14.2
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
yarl==1.9.4
